+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Set minimal supported Gradle version to 6.7
 - Fix for `getHeaderField("Location") must not be null` [#960](../../issues/960)
 - Added `BuildFeature` feature flags
+- `instrumentCode` task – changes in Kotlin code no longer rebuild the plugin [#959](../../issues/959)
 
 ## 1.5.3
 - Updated dependencies marked as vulnerable [#459](../../issues/459)